api/pyproject.toml
[project]
name = "capy-privacy-api"
version = "0.1.0"
description = ""
authors = [
    {name = "gael-soude",email = "capy.security@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "peewee (>=3.18.3,<4.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry]
packages = [{include = "api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
